move variable form into the general components folder
