+++ ui/src/components/VariableForm/FileUpload/__tests__/utils.test.ts
